pyproject.toml
[project]
name = "powerthon-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.0",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "appnope==0.1.4 ; sys_platform == 'darwin'",
    "asttokens==3.0.0",
    "attrs==25.4.0",
    "certifi==2025.10.5",
    "cffi==2.0.0 ; implementation_name == 'pypy'",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "colorama==0.4.6",
    "comm==0.2.3",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.17",
    "decorator==5.2.1",
    "distro==1.9.0",
    "executing==2.2.1",
    "fastapi==0.119.0",
    "frozenlist==1.8.0",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "idna==3.11",
    "ipykernel==7.0.1",
    "ipython==9.6.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jiter==0.11.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "langchain==0.3.27",
    "langchain-community==0.3.31",
    "langchain-core==0.3.79",
    "langchain-ollama==0.3.10",
    "langchain-openai==0.3.35",
    "langchain-text-splitters==0.3.11",
    "langsmith==0.4.37",
    "marshmallow==3.26.1",
    "matplotlib>=3.10.7",
    "matplotlib-inline==0.1.7",
    "multidict==6.7.0",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "numpy==2.3.4",
    "ollama==0.6.0",
    "openai==2.3.0",
    "orjson==3.11.3",
    "packaging==25.0",
    "pandas==2.3.3",
    "parso==0.8.5",
    "pexpect==4.9.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'",
    "platformdirs==4.5.0",
    "prompt-toolkit==3.0.52",
    "propcache==0.4.1",
    "psutil==7.1.0",
    "ptyprocess==0.7.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'",
    "pure-eval==0.2.3",
    "pycparser==2.23 ; implementation_name == 'pypy'",
    "pydantic==2.12.2",
    "pydantic-core==2.41.4",
    "pydantic-settings==2.11.0",
    "pygments==2.19.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pywin32==311",
    "pyyaml==6.0.3",
    "pyzmq==27.1.0",
    "referencing==0.37.0",
    "regex==2025.9.18",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rpds-py==0.27.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.44",
    "stack-data==0.6.3",
    "starlette==0.48.0",
    "tenacity==9.1.2",
    "tiktoken==0.12.0",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.37.0",
    "wcwidth==0.2.14",
    "yarl==1.22.0",
    "zstandard==0.25.0",
]
